pylint: locally suppress misplaced-comparison-constant
When comparing variables with constants, pylint likes for the constant
to appear on the right hand side. When testing our constants, we
deliberately test comparisons both ways round. It may be the case that
these tests are ultimately equivalent under the hood, but we keep both
versions to be sure.